Add Serial.printHex, support for disabling SerialEvent
Serial.printHex() now exists. It's non-standard, but oh so convenient. Lately I've been pasting blocks of code like that into almost all of my sketches when debugging, and that was bound to one port. I may move this to a library? Not sure how I could make it work on all ports in a natural way though. Takes 8, 16, and 32-bit values, for the 16 and 32, there's a third argument to swap the endianness (byte order) - computer brains put the low digit first, human brains put the high digit first. Finally, you can pass it a pointer to a uint8_t or uint16_t (byte or unsigned int), tell it how many entries you want, and what character you want to separate them with.

This also adds the option (not yet exposed) to disable SerialEvent, which should never have existed.