Adapt SPI library to use either SPI peripheral
It will either use SPI0 or SPI1. This was a... rather large task! The resulting library should be good to go for all AVRxt parts. And end now reverses begin (turning off the outputs it turned on... but you can now pass the included defines pointing to the other SPI port (these are listed in the readme now!)
#35, #32