Dynamic PWM
DYNAMIC PWM PIN MAPPING!

digitalPinToTimerNow(pin) - returns the the code for timer (a byte constant, see Arduino.h) that is currently controlling PWM, or NOT_ON_TIMER if there isn't one. This first checks if TCA0 is, then if TCA1, and then uses digitalPinToTimer() (the old one) which looks up in the table,returning any TCD or TCB associations. For TCB's we also check CTRLB and make sure that the timer is in PWM mode, lest analogWrite charge ahead blindly and break a timer configuration.
Adapt pins_arduino for the variants to deal with this as well.