eeek, bad output bug in comparator and zcd
This was nasty. They all share 2 output pins (also shared with zcd). If i have one that is set to have output, and others have it disabled because they are event or interrupt generators, if I init the output-having one first first, the others turn pin back to input! That would have even broken code that configured those pins for totally different purpose. Never set a pin you are not using to be an input- you don't know what sketch already did with it on purpose. Only manipulate anything about pins being used.

The example set by official API is that things that need to be output should be set that way for you, but library and core code doesn't have to clean up afterwards if it gets disabled later.