remove dissassembler scripts, and wire1
It was found that the dissassembler scripts were unneccessary based on ArminJo's repo. Also stole his idea on memory map too.  Meanwhile, Wire1 doesn[t even compile, and never did what it was supposed to. Out it goes - I don't have the skill to make what needs to be done for a Wire1 library happen