fix tones of bugs in ton()
Correct large number of tone() bugs.
 Tone could leave the pin high.
A high pitch and long duration could result in much shorter tones than intended.
Frequency of 0 would leave the timer running and generating interrupts in the background for the requested duration (now it stops output if it's on the current pin, otherwise does nothing). while doing nothing.
No attempt was made to handle invalid pins: it would get null pointer to port struct.... and proceed to use it without testing if it was valid, or if the bitmask was valid. Now we test for valid bitmask before requesting the port struct.
Frequencies over 32768 would overflow an intermediate when called with duration and the duration would be very short.
Long durations can also overflow the intermediate. The intermediate is found as `2 * frequency * duration` and both multiplications can overflow (first one is as unsigned int, second as signed long.) Since the result is divided by 1000, remove 2* and change to /500 doubled the limit and removed the problem above 32768 Hz.
At least one variable that was volatile didnt need to be, and one that should be was not, and interrupts were left off for far longer than they needed to be. We also dont awkwardly recalculate everything if the frequency hasnt changed since the last time tone() was called.