minor stuff
pinMode compile-errors when passed constants that aren't valid modes. Unlike digitalWrite, where values that are neiuther HIGH nor LOW are routinely used in place of HIGH when convenient, that sort of thing is never done with pinMode (not only is the non-binary nature of it less conducive to that, the behavior with things that are neither HIGH nor LOW varies between cores.
imoroive digititalPinToTimerNow() for TCD (though code still commented out, until either DD or new silkicon rev of DA/DB becomes available.)
Improve comments on both files.