Docs + wire formatting improvements
Realized that we were trying too hard to avoid compare being used in micros. This method is far more tolerant of calls with interrupts disabled. We were previously assuming, in effect, that if the timer count was more than a few ticks after it overflowed, that the flag got set after we read the count but before the flags. That seems very unlikely since we read them on consecutive clock cycles (it's more of an issue with TCD, where we don't get that luxury.. Even what this implements is more cautious than it needs to be.