Rework the fix for #226
I wanted to keep the new errors in by default - I think as long as success returns 0, nearly all code in the wild will work. and the small amount of code that does break got broken a few months back when the timeout code was added. Let's see how this goes over and the author should be getting a message ;-)