dd spi pinmux correction
the SPI pinswap defines are inconsistent with DS40002413A-page 155:

pinswap 0x03 is reserved on 14-pin and 20-pin parts. its pins (A0,A1) do not occur in the table. They must have been copied from the 28/32 pin where they are correct.

and PIN_SPI_SS_PINSWAP_6 should be PIN_PF7 (UPDI) not NOT_A_PIN - or does defining it as NOT_A_PIN behave better with UPDI not fused to GPIO? In which case perhaps a comment (assuming an #ifdef against whatever drives that fuse option is impossible). But the 32-pin part already had PF7 so putting it in the other 3.

I added the defaults for PIN_SPI_X to NOT_A_PIN for consistency with TWI0, below where they were missing.

Head of tree is broken for me, this is completely untested (ran across it when I noticed a discrepancy between the mappings generated from parsing pins_arduino and the datasheet extracted pinmap).