switch over to the atpack-derived diagrams
there were a few (CCL3) signals that were getting dropped from the datasheet derived since they were a signal type that was normally in the portmux. There were also some bogus signals that were errors in the data sheet that I have logic to clean up in the atpack derived diagrams.

The change to file naming reflects the chip from which the data was pulled. The diagrams continue to be titled generically.