Add Azduino5 toolchain package
Hosted in github because I would be eaten alive by the bandwith costs of CI testing, which installs the package for each of 72 VMs every time a commit changes code, It's in everyone's interest for that traffic to remain internal to github, but the key issue is that outside hosting like is used for the other files is not free. As they only get installed once per system that has either core installed, outside of CI testing, the traffic isn't costly. But it is prohibitively expensive if the CI tests are run that way.