Fix #365
The initialization loop for the TCBs had the timer incremented inside the body of the if statement. So when it hit the timer used for millis, it would not increment, and sat there until the end of time. This, combined with the confusing thicket of #ifdefs around the startup code and an expectation that the problem was triggered by the interrupt (assumption of a complex problem obscured the existence of a simple problem.