Codespell.
Also, "shat" absolutely is a word. It is the past tense of "shit" when used as a verb. How could you talk about assembly. particularly inline, without a proper vocabulary of profanity in different tenses? If you shit on a register, you've gotta tell the compiler which register you shat on!