Update redirecting URLs in bundled library metadata
The `url` field of the `library.properties` metadata file points the user to where they might find more information or
assistance for the library.

This platform includes bundled libraries which are variants of the standardized libraries distributed by Arduino. The
`url` field in the metadata of these libraries points to the reference on the arduino.cc website, as that information is
also applicable to this platform's variants of the libraries.

Due to restructuring of the arduino.cc website, the targeted reference pages have moved to a different URL. Arduino set
up redirects from the previous URLs to the new one, but redirects have a tendency to break or no longer lead to the
intended target over time so it is best to avoid relying on them. In addition, the use of redirecting URLs seems to make
it more likely that automated link checks will fail due to an HTTP 403 ("Forbidden") response code, even though the URL
does work as expected when a human loads it in their browser.